refactor: scope system status ui to generation
