+++ app/frontend/src/composables/system/index.ts
@@ -1,3 +1,2 @@
-export * from './useSystemStatus';
 export * from './useAdminMetrics';
