+++ app/frontend/src/features/generation/components/index.ts
@@ -4,3 +4,9 @@ export { default as GenerationResultsGallery } from './GenerationResultsGallery.
 export { default as GenerationStudio } from './GenerationStudio.vue';
 export { default as GenerationSystemStatusCard } from './GenerationSystemStatusCard.vue';
 export { default as JobQueue } from './JobQueue.vue';
+export { default as SystemAdminStatusCard } from './system/SystemAdminStatusCard.vue';
+export { default as SystemStatusCard } from './system/SystemStatusCard.vue';
+export { default as SystemStatusCardDetailed } from './system/SystemStatusCardDetailed.vue';
+export { default as SystemStatusCardSimple } from './system/SystemStatusCardSimple.vue';
+export { default as SystemStatusCardSkeleton } from './system/SystemStatusCardSkeleton.vue';
+export { default as SystemStatusPanel } from './system/SystemStatusPanel.vue';
