+++ app/frontend/src/features/generation/components/system/SystemAdminStatusCard.vue
