+++ app/frontend/src/features/generation/components/system/SystemStatusCard.vue
@@ -55,7 +55,7 @@
 <script setup lang="ts">
 import { computed, ref, watch } from 'vue';
 
-import { useSystemStatus } from '@/composables/system';
+import { useSystemStatus } from '../../composables/useSystemStatus';
 import SystemStatusCardDetailed from './SystemStatusCardDetailed.vue';
 import SystemStatusCardSimple from './SystemStatusCardSimple.vue';
 import SystemStatusCardSkeleton from './SystemStatusCardSkeleton.vue';
