+++ app/frontend/src/features/generation/components/system/SystemStatusCardDetailed.vue
