+++ app/frontend/src/features/generation/components/system/SystemStatusCardSimple.vue
