+++ app/frontend/src/features/generation/components/system/SystemStatusCardSkeleton.vue
