+++ app/frontend/src/features/generation/components/system/SystemStatusPanel.vue
