+++ app/frontend/src/features/generation/composables/index.ts
@@ -14,4 +14,5 @@ export * from './useJobQueue';
 export * from './useJobQueueActions';
 export * from './useJobQueuePolling';
 export * from './useJobQueueTransport';
+export * from './useSystemStatus';
