+++ app/frontend/src/views/ImportExportView.vue
@@ -22,9 +22,8 @@
 import { defineAsyncComponent, ref } from 'vue';
 import { RouterLink } from 'vue-router';
 
-import { JobQueue } from '@/features/generation';
+import { JobQueue, SystemStatusPanel } from '@/features/generation';
 import PageHeader from '@/components/layout/PageHeader.vue';
-import SystemStatusPanel from '@/components/system/SystemStatusPanel.vue';
 import ImportExportSkeleton from '@/components/import-export/ImportExportSkeleton.vue';
 
 const LazyImportExportContainer = defineAsyncComponent({
