+++ app/frontend/src/views/RecommendationsView.vue
@@ -26,5 +26,5 @@ import { RouterLink } from 'vue-router';
 import { GenerationHistory } from '@/features/history';
 import PageHeader from '@/components/layout/PageHeader.vue';
 import { RecommendationsPanel } from '@/features/recommendations';
-import SystemStatusCard from '@/components/system/SystemStatusCard.vue';
+import { SystemStatusCard } from '@/features/generation';
 </script>
